+++ R/report.R
@@ -1,10 +1,11 @@
 #' Generate report on Revision Analysis
 #'
-#' @param rslt an object of class `"rjd3rev_vintages"` which is the output
+#' @param rslt an object of class `"rjd3rev_rslts"` which is the output
 #'             of the function `revision_analysis()`
 #' @param output_file path or name of the output file containing the report
 #' @param output_dir path of the dir containing the output file (Optional)
-#' @param output_format either an HTML document (default) or a PDF document
+#' @param output_format either an HTML document (default), a PDF document or a
+#' Word document
 #' @param plot_revisions Boolean. Default is FALSE meaning that a plot with the
 #'   revisions will not be added to the report.
 #' @param open_report Boolean. Default is TRUE meaning that the report will
@@ -20,25 +21,16 @@
 #' @examples
 #'
 #' ## Simulated data
-#' period_range <- seq(as.Date('2011-01-01'),as.Date('2020-10-01'),by='quarter')
-#' qtr <- (as.numeric(substr(period_range,6,7))+2)/3
-#' time_period <- rep(paste0(format(period_range, "%Y"), "Q", qtr),5)
-#' np <- length(period_range)
-#' rev_date <- c(rep("2021-06-30",np), rep("2021-12-31",np), rep("2022-06-30",np),
-#'             rep("2022-12-31",np), rep("2023-06-30",np))
-#' set.seed(1)
-#' xt <- cumsum(sample(rnorm(1000,0,1), np, TRUE))
-#' rev <- rnorm(np*4,0,.1)
-#' obs_values <- xt
-#' for(i in 1:4) {
-#'   xt <- xt+rev[(1+(i-1)*np):(i*np)]
-#'   obs_values <- c(obs_values,xt)
-#' }
-#' df <- data.frame(rev_date, time_period, obs_values)
+#' df_long <- simulate_long(
+#'     n_period = 10L * 4L,
+#'     n_revision = 5L,
+#'     periodicity = 4L,
+#'     start_period = as.Date("2010-01-01")
+#' )
 #'
 #' ## Make analysis and generate the report
 #'
-#' vintages <- create_vintages(df, periodicity = 4L, type = "long")
+#' vintages <- create_vintages(df_long, periodicity = 4L, type = "long")
 #' rslt <- revision_analysis(vintages, view = "diagonal")
 #'
 #' \dontrun{
@@ -51,13 +43,14 @@
 #' )
 #' }
 #'
-render_report <- function(rslt,
-                          output_file,
-                          output_dir,
-                          output_format = c("html_document", "pdf_document"),
-                          plot_revisions = FALSE,
-                          open_report = TRUE,
-                          ...) {
+render_report <- function(
+        rslt,
+        output_file,
+        output_dir,
+        output_format = c("html_document", "pdf_document", "word_document"),
+        plot_revisions = FALSE,
+        open_report = TRUE,
+        ...) {
 
     # Check input
     checkmate::assert_class(rslt, "rjd3rev_rslts")
@@ -81,6 +74,8 @@ render_report <- function(rslt,
             ext <- "html"
         } else if (output_format == "pdf_document") {
             ext <- "pdf"
+        } else if (output_format == "word_document") {
+            ext <- "docx"
         }
     }
     output_file <- tools::file_path_sans_ext(output_file)
@@ -95,6 +90,7 @@ render_report <- function(rslt,
     }
 
     e <- list2env(list(
+        rslt = rslt,
         descriptive_statistics = rslt$descriptive.statistics,
         main_results = rslt$summary,
         add_plot = plot_revisions,